+++ src/components/UI/atoms/sorts/Sorts.jsx
@@ -5,6 +5,9 @@ import { responsive } from '../../../../style/responsive';
 import MobileSort from '../../../molecules/mobileSort/MobileSort';
 
 const Sorts = ({ className, sortState, setSortState, setPageState }) => {
+  // sortState -> 현재 클릭된 sort 상태를 props로 받는다 ex) 최신순, 인기순...
+  // setSortState -> sortState를 설정하는 함수를 props로 받는다.
+  // setPageState -> pageState를 설정하는 함수를 props로 받는다.
   // onClick 이벤트에 각각 데이터 불러오는 상태 함수를 props로 받아와햔다.
   const [mobileCurrentSort, setMobileCurrentSort] = useState('최신순');
   const [mobileSortIsClicked, setMobileSortIsClicked] = useState(false);
